+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.7] - 2025-11-12
+
+- Initial public release of `ms_single_multi_select_dropdown`
+- Supports single and multi-select modes
+- Icon and icon color change
+- Background color changing option
+
 ## [0.0.6] - 2025-11-12
 
 - Initial public release of `ms_single_multi_select_dropdown`